requirements.txt
streamlit_extras
polars
nltk
streamlit_option_menu
plotly
spacy
en-core-web-md